pyproject.toml
[project]
name = "customer-service-agent-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.5.0",
    "google-api-python-client>=2.174.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "poetry>=2.1.3",
    "pymongo>=4.13.2",
    "requests>=2.32.4",
]
